update raise hand handler

+++ backend/src/handlers/onUserDisconnect.ts
@@ -10,9 +10,16 @@ export const onUserDisconnect: MessageHandler<IUserDisconnected> = async (
 ) => {
   const { user } = data;
 
+  const stream = await ParticipantService.getCurrentStreamFor(user);
+
+  if (!stream) {
+    return;
+  }
+
   await Promise.all([
     FeedsCacheService.removeServedStreams(user),
     FeedService.removeCandidateByOwner(user),
+    ParticipantService.removeParticipant(user, stream),
     ParticipantService.broadcastParticipantLeft(user),
   ]);
 };
